New constructor in Pengine, fixed missile.
Added new constructor to Pengine - Pengine(PVector, float, float,
ColorTransition, BOOLEAN). Boolean argument is true when you want
particles to emit light. Its quite nice effect.
I'm using this constructor in block -> so on explosion of block,
particles emit light. I don't know about it, it looks nice, and the
range of the particle light is small so it's not OP :D You can turn it
off, if you want. :)